# docs/docs-site/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Community Feed API",
    "description": "Community-governed Bluesky feed generator where subscribers democratically vote on ranking algorithm weights. Built on AT Protocol.\n\n## Authentication\n- **Governance endpoints** require a session cookie or bearer token from `POST /api/governance/auth/login`.\n- **Admin endpoints** additionally require the caller's DID to be in the `BOT_ADMIN_DIDS` allowlist.\n- **Feed endpoints** are public (called by the Bluesky app). Auth is optional for subscriber tracking.\n- **Transparency endpoints** are public and unauthenticated.",
    "version": "1.1.0",
    "contact": {
      "name": "Community Feed",
      "url": "https://github.com/andrewnordstrom-eng/bluesky-community-feed"
    },
    "license": {
      "name": "MIT"
    }
  },
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "governance_session",
        "description": "Session cookie from POST /api/governance/auth/login"
      },
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Bearer token from POST /api/governance/auth/login (for CLI/MCP)"
      }
    },
...
  },
  "paths": {
...
    "/api/governance/research-consent": {
      "get": {
        "summary": "Get research consent status",
        "tags": [
          "Governance"
        ],
        "description": "Returns the authenticated user's current research consent status. Research consent is separate from TOS — declining does not restrict governance access.",
        "security": [
          {
            "cookieAuth": []
          },
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "consent": {
                      "type": "boolean",
                      "nullable": true,
                      "description": "Current consent status"
                    },
                    "consentedAt": {
                      "type": "string",
                      "format": "date-time",
                      "nullable": true
                    },
                    "consentVersion": {
                      "type": "string",
                      "nullable": true
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "description": "Error code (e.g. \"ValidationError\", \"NotFound\")"
                    },
                    "message": {
                      "type": "string",
                      "description": "Human-readable error message"
                    },
                    "correlationId": {
                      "type": "string",
                      "description": "Request correlation ID for debugging"
                    }
                  },
                  "required": [
                    "error",
                    "message"
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "description": "Error code (e.g. \"ValidationError\", \"NotFound\")"
                    },
                    "message": {
                      "type": "string",
                      "description": "Human-readable error message"
                    },
                    "correlationId": {
                      "type": "string",
                      "description": "Request correlation ID for debugging"
                    }
                  },
                  "required": [
                    "error",
                    "message"
                  ]
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Record research consent",
        "tags": [
          "Governance"
        ],
        "description": "Record or update the user's research consent decision. Consent changes are logged to the audit trail.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {}
            }
          }
        },
        "security": [
          {
            "cookieAuth": []
          },
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    }
                  },
                  "required": [
                    "success"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "description": "Error code (e.g. \"ValidationError\", \"NotFound\")"
                    },
                    "message": {
                      "type": "string",
                      "description": "Human-readable error message"
                    },
                    "correlationId": {
                      "type": "string",
                      "description": "Request correlation ID for debugging"
                    }
                  },
                  "required": [
                    "error",
                    "message"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "description": "Error code (e.g. \"ValidationError\", \"NotFound\")"
                    },
                    "message": {
                      "type": "string",
                      "description": "Human-readable error message"
                    },
                    "correlationId": {
                      "type": "string",
                      "description": "Request correlation ID for debugging"
                    }
                  },
                  "required": [
                    "error",
                    "message"
                  ]
                }
              }
            }
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "https://feed.yourdomain.com",
      "description": "Production"
    }
  ],
...
}